Restructure:
1. restructure project with pyckage-ccoekicutter
2. fix issue, that package is incompatible with python version bigger than 3.8.6, because the changed interface of function _get_mixins_ in EnumMeta class in commit python/cpython@5efb1a7#diff-e5dd6e444e4e6c2cdeb8edf271d057f9da52f7b406dcd8b655dc396fb013bd03
