+++ .github/workflows/test.yml
@@ -18,15 +18,32 @@ on:
 
 # A workflow run is made up of one or more jobs that can run sequentially or in parallel
 jobs:
+  # This workflow contains two jobs "linting" and "test"
+  linting:
+    # The type of runner that the job will run on
+    runs-on: ubuntu-latest
+    # Steps represent a sequence of tasks that will be executed as part of the job
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.8'
+    - name: Run flake8 check
+      run: make flake8
+    - name: Run mypy check
+      run: make mypy
+    - name: Run bandit check
+      run: make bandit
   # This workflow contains a single job called "build"
   test:
+    needs: linting
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
     strategy:
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: [3.6, 3.7, 3.8]
-
+        python-version: [3.6, 3.7, 3.8.5, 3.8, 3.9]
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
     # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
@@ -35,41 +52,34 @@ jobs:
       uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
-
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r test-requirements.txt
-
+        pip install -r requirements/dev.txt
+        pip install -e .
     - name: Test with pytest
       run: |
         pytest
         coverage xml
-
     - name: Codecov Action
       uses: Atrox/codecov-action@v0.1.3
-
+  # This workflow contains a single job called "build"
   build:
+    needs: test
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [3.6, 3.7, 3.8]
-
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - uses: actions/checkout@v2
-      - name: Set up Python ${{ matrix.python-version }}
+      - name: Set up Python 3.8
         uses: actions/setup-python@v2
         with:
-          python-version: ${{ matrix.python-version }}
-
+          python-version: 3.8
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install setuptools wheel twine
-
-      - name: Build
+      - name: Build and check package with twine
         run: |
           python setup.py sdist bdist_wheel
+          twine check dist/*.whl
