+++ .gitignore
@@ -1,14 +1,57 @@
-.*
-!.gitignore
-__pycache__
+# Temporary and binary files
+*~
+*.py[cod]
+*.so
+*.cfg
+!setup.cfg
+*.orig
+*.log
+*.pot
+*/__pycache__/*
+.cache/*
+.*_cache/*
+.*.swp
+*/.ipynb_checkpoints/*
 *.pyc
+
+# Project files
+.ropeproject
+.project
+.pydevproject
+.settings
+.idea
+
+# Package files
+*.egg
+*.eggs/
+.installed.cfg
 *.egg-info
-build
-_static
-_templates
-dist
-!.readthedocs.yml
-!.travis.yml
-!.pyup.yml
-!.github
+
+# Unittest and coverage
+htmlcov/*
+.coverage
+.tox
+junit.xml
 coverage.xml
+
+# Build and docs folder/files
+build/*
+dist/*
+sdist/*
+docs/api/*
+docs/_build/*
+docs/build/*
+cover/*
+MANIFEST
+.mypy_cache/*
+
+# Unused code
+backup/*
+
+# macOS Desktop Services Store
+.DS_Store
+
+# plots
+*.png
+
+venv
