--- docs/index.rst
@@ -1,99 +1,0 @@
-.. named_enum documentation master file, created by
-   sphinx-quickstart on Wed Dec  5 13:54:39 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-named_enum's documentation!
-======================================
-
-Introduction
-------------
-
-This package provides several enumeration classes, which extends the default **Enum** class with various functionalities. For each enumeration class, its enumeration item's value is a customised tuple type generated by **namedtuple** from **collections** package.
-
-If you like it, please |start|_ it in |github|_.
-
-.. |start| image:: https://img.icons8.com/material-rounded/26/000000/star.png
-    :alt: star
-
-.. |github| image:: https://img.icons8.com/small/26/000000/github.png
-    :alt: github
-
-.. _github: https://github.com/KnightConan/named_enum
-
-.. _start: https://github.com/KnightConan/named_enum
-
-Status
-------
-.. license badge
-.. image:: https://img.shields.io/pypi/l/named-enum.svg
-    :target: https://pypi.python.org/pypi/named-enum/
-
-.. readthedocs badge
-.. image:: https://readthedocs.org/projects/named-enum/badge/?version=latest
-    :target: https://named-enum.readthedocs.io/en/latest/?badge=latest
-    :alt: Documentation Status
-
-.. actions building badge
-.. image:: https://github.com/KnightConan/named_enum/workflows/Unit%20Test%20&%20Build%20Test/badge.svg
-    :target: https://github.com/KnightConan/named_enum/actions
-
-.. pypi version badge
-.. image:: https://img.shields.io/pypi/v/named-enum.svg
-    :target: https://pypi.python.org/pypi/named-enum/
-
-.. development status from pypi
-.. image:: https://img.shields.io/pypi/status/named-enum.svg
-    :target: https://pypi.python.org/pypi/named-enum/
-
-.. python version badge from PyPI
-.. image:: https://img.shields.io/pypi/pyversions/named-enum.svg
-    :target: https://pypi.python.org/pypi/named-enum/
-    :alt: Python 3.6 | Python 3.7
-
-.. pypi format
-.. image:: https://img.shields.io/pypi/format/named-enum.svg
-    :target: https://badge.fury.io/py/named-enum
-
-.. codecov badge
-.. image:: https://codecov.io/gh/KnightConan/named_enum/branch/master/graph/badge.svg
-    :target: https://codecov.io/gh/KnightConan/named_enum
-
-.. pyup badge
-.. image:: https://pyup.io/repos/github/KnightConan/named_enum/shield.svg
-    :target: https://pyup.io/repos/github/KnightConan/named_enum/
-    :alt: Updates
-
-.. download statistics badge
-.. image:: https://pepy.tech/badge/named-enum
-    :target: https://pepy.tech/project/named-enum
-
-.. send message to the author
-.. image:: https://img.shields.io/badge/saythanks.io-☼-1EAEDB.svg
-    :target: https://saythanks.io/to/KnightConan
-
-Developers
-----------
-**Author**: `Zhiwei Zhang <https://github.com/KnightConan>`_
-
-**Email**: zhiwei.zhang@outlook.de
-
-**Github**: `named_enum <https://github.com/KnightConan/named_enum>`_
-
-License
--------
-`MIT <https://github.com/KnightConan/named_enum/blob/master/LICENSE>`_
-
-Acknowledgement
----------------
-- `Cristian Alfonso González Mora <https://github.com/cagonza6/>`_ for the inspiration of this project.
-
-
-.. toctree::
-   :maxdepth: 2
-   :hidden:
-
-   setup
-   quick_start
-   source_code
-   release_notes
