--- docs/setup.rst
@@ -1,15 +1,0 @@
-============
-Installation
-============
-
-This package is deployed in `PyPI <https://pypi.org/project/named-enum/>`_,
-you can install it through the command:
-
-``pip install named-enum``
-
-running test
-````````````
-
-    .. code-block:: console
-
-        python setup.py test
