+++ docs/source/01_about.rst
@@ -1,0 +1,1 @@
+.. include:: ../../README.rst
