+++ docs/source/03_release_notes.rst
@@ -1,4 +1,3 @@
-=============
 Release Notes
 =============
