+++ docs/source/04_authors.rst
@@ -1,0 +1,7 @@
+Authors
+=======
+
+..
+    Format: `Name <mailto:email>`_ - *Role/Responsibility* - `email <mailto:email?subject=[GitHub]Named%20Enum>`_
+
+* `Zhiwei Zhang <https://github.com/zhiwei2017>`_ - *Maintainer* - `zhiwei2017@gmail.com <mailto:zhiwei2017@gmail.com?subject=[GitHub]Named%20Enum>`_
