+++ docs/source/05_contributing.rst
@@ -1,0 +1,1 @@
+.. include:: ../../CONTRIBUTING.rst
