+++ docs/source/index.rst
@@ -1,0 +1,22 @@
+===========================================
+Welcome to the documentation of Named Enum!
+===========================================
+
+Index
+-----
+.. toctree::
+   :maxdepth: 1
+
+   01_about
+   02_source
+   03_release_notes
+   04_authors
+   05_contributing
+
+   
+Indices and tables
+------------------
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
