+++ named_enum/version.py
@@ -1,0 +1,2 @@
+"""Define the version of the web service"""
+__version__ = "1.0.2"
