+++ requirements/doc.txt
@@ -1,0 +1,3 @@
+-r base.txt
+sphinx~=3.3
+sphinx_rtd_theme~=0.5
