+++ setup.cfg
@@ -1,10 +1,19 @@
+[bumpversion]
+current_version = 1.0.2
+commit = True
+tag = True
+
 [flake8]
+ignore = F401,E501
+select = B,C,E,F,W,T4
+max-line-length = 99
+max-complexity = 18
 exclude =
-    .tox,
     .git,
-    __pycache__,
-    build,
-    dist
+    __pycache__
+
+[mypy]
+ignore_missing_imports = false
 
 [tool:pytest]
 doctest_encoding = latin1
@@ -13,7 +22,8 @@ addopts =
   --verbose
   --doctest-modules
   --doctest-glob=tests/*.txt
-  --cov=named_enum --cov-report=xml tests/
+  --cov=./named_enum
+  --cov-report=xml
 
 [aliases]
 test=pytest
