--- test-requirements.txt
@@ -1,5 +1,0 @@
-pytest==5.4.3
-pytest-cov==2.10.0
-pytest-runner==5.2
-codecov==2.1.7
-coverage==5.1
