+++ tests/test_version.py
@@ -1,0 +1,6 @@
+from named_enum.version import __version__
+
+
+class TestVersion:
+    def test_version(self):
+        assert isinstance(__version__, str)
