+++ tox.ini
@@ -1,0 +1,27 @@
+[tox]
+envlist = flake8, mypy, bandit, python3.6, python3.7, python3.8, python3.9
+skipsdist = true
+
+[testenv]
+platform = linux2|darwin
+setenv =
+    PYTHONPATH = {toxinidir}
+deps =
+    -r{toxinidir}/requirements/dev.txt
+    -e .
+commands =
+    pytest
+
+[testenv:flake8]
+deps = flake8
+commands = flake8 named_enum
+
+[testenv:mypy]
+deps =
+    -e .
+    mypy
+commands=mypy named_enum
+
+[testenv:bandit]
+deps = bandit
+commands = bandit -r named_enum
